+++ backend/src/main/java/com/foodly/backend/entity/Ingredient.java
@@ -2,6 +2,8 @@ package com.foodly.backend.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -21,14 +23,14 @@ public class Ingredient {
     private String name;
 
     @Column(name = "calories_per_100g")
-    private double caloriesPer100g;
+    private BigDecimal caloriesPer100g;
 
     @Column(name = "proteins_per_100g")
-    private double proteinsPer100g;
+    private BigDecimal proteinsPer100g;
 
     @Column(name = "fats_per_100g")
-    private double fatsPer100g;
+    private BigDecimal fatsPer100g;
 
     @Column(name = "carbs_per_100g")
-    private double carbsPer100g;
+    private BigDecimal carbsPer100g;
 }
